narrow the inclusions from Rumali's branch
to just the comparison helper functions
